# docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /api/v1/product:
    get:
      tags:
      - product-controller
      summary: Получить страницу объектов
      operationId: productGetAll
      parameters:
      - name: pageNumber
        in: query
        description: Номер страницы
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: Размер страницы
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: Успешно получено
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageDtoProductDto"
    put:
      tags:
      - product-controller
      summary: "Обновить объект полностью, кроме его ID"
      operationId: productUpdate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductDto"
        required: true
      responses:
        "200":
          description: Успешно обновлено
    post:
      tags:
      - product-controller
      summary: Создать объект и получить в ответ его ID
      operationId: productCreate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductDto"
        required: true
      responses:
        "200":
          description: Успешно создано
          content:
            '*/*':
              schema:
                type: integer
                format: int32
...
  /api/v1/product/unitOfMeasure/productWhereUnitOfMeasureIsSet:
    post:
      tags:
      - product-operations-controller
      operationId: productoperationsGetUnitOfMeasureProductWhereUnitOfMeasureIsSet
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                enum:
                - METERS
                - CENTIMETERS
                - LITERS
                - GRAMS
                - MILLIGRAMS
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductDto"
  /api/v1/product/price/reduce:
    post:
      tags:
      - product-operations-controller
      operationId: productoperationsReducePrice
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceReduceDto"
        required: true
      responses:
        "200":
          description: OK
...
  /api/v1/product/rating/countOfProductWhereRatingGreaterThan/{rating}:
    get:
      tags:
      - product-operations-controller
      operationId: productoperationsGetRatingCountOfProductWhereRatingGreaterThat
      parameters:
      - name: rating
        in: path
        required: true
        schema:
          type: number
          format: double
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int64
...
components:
  schemas:
    AddressDto:
      type: object
      description: объект
      properties:
        id:
          type: integer
          format: int32
        street:
          type: string
        zipCode:
          type: string
        town:
          $ref: "#/components/schemas/LocationDto"
      required:
      - id
    CoordinatesDto:
      type: object
      description: объект
      properties:
        id:
          type: integer
          format: int32
        x:
          type: number
          format: double
        "y":
          type: number
          format: float
      required:
      - id
    LocationDto:
      type: object
      description: объект
      properties:
        id:
          type: integer
          format: int32
        x:
          type: number
          format: float
        "y":
          type: integer
          format: int64
        z:
          type: integer
          format: int64
      required:
      - id
    OrganizationDto:
      type: object
      description: объект
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        officialAddress:
          $ref: "#/components/schemas/AddressDto"
        annualTurnover:
          type: integer
          format: int32
        employeesCount:
          type: integer
          format: int64
        fullName:
          type: string
        rating:
          type: integer
          format: int32
        type:
          type: string
          enum:
          - COMMERCIAL
          - PUBLIC
          - TRUST
          - OPEN_JOINT_STOCK_COMPANY
      required:
      - id
    PersonDto:
      type: object
      description: объект
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        eyeColor:
          type: string
          enum:
          - BLUE
          - YELLOW
          - ORANGE
        hairColor:
          type: string
          enum:
          - BLUE
          - YELLOW
          - ORANGE
        location:
          $ref: "#/components/schemas/LocationDto"
        birthday:
          type: string
          format: date-time
        nationality:
          type: string
          enum:
          - USA
          - CHINA
          - VATICAN
          - THAILAND
          - NORTH_KOREA
      required:
      - id
    ProductDto:
      type: object
      description: объект
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        coordinates:
          $ref: "#/components/schemas/CoordinatesDto"
        creationDate:
          type: string
          format: date-time
        unitOfMeasure:
          type: string
          enum:
          - METERS
          - CENTIMETERS
          - LITERS
          - GRAMS
          - MILLIGRAMS
        manufacturer:
          $ref: "#/components/schemas/OrganizationDto"
        price:
          type: number
          format: double
        manufactureCost:
          type: integer
          format: int32
        rating:
          type: number
          format: float
        partNumber:
          type: string
        owner:
          $ref: "#/components/schemas/PersonDto"
      required:
      - id
    PriceReduceDto:
      type: object
      properties:
        percent:
          type: number
          format: double
          maximum: 100
          minimum: 0
    PageDtoProductDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ProductDto"
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalSize:
          type: integer
          format: int64
...